+++ src/main/java/com/bulmansoda/map_community/service/MarkerService.java
@@ -3,8 +3,10 @@ package com.bulmansoda.map_community.service;
 import com.bulmansoda.map_community.dto.ai.GptRequest;
 import com.bulmansoda.map_community.dto.ai.GptResponse;
 import com.bulmansoda.map_community.dto.marker_service.CreateMarkerRequest;
+import com.bulmansoda.map_community.exception.AuthorizationException;
 import com.bulmansoda.map_community.exception.CenterMarkerNotFoundException;
 import com.bulmansoda.map_community.exception.MarkerNotFoundException;
+import com.bulmansoda.map_community.exception.UserNotFoundException;
 import com.bulmansoda.map_community.model.CenterMarker;
 import com.bulmansoda.map_community.model.Marker;
 import com.bulmansoda.map_community.model.User;
@@ -34,12 +36,12 @@ public class MarkerService {
         this.aiClusteringService = aiClusteringService;
     }
 
-    public long createMarker(CreateMarkerRequest request) {
+    public long createMarker(CreateMarkerRequest request, Long userId) {
         Marker marker = new Marker();
         marker.setLatitude(request.getLatitude());
         marker.setLongitude(request.getLongitude());
-        User user = userRepository.findById(request.getUserId())
-                .orElseThrow(() -> new MarkerNotFoundException(request.getUserId()));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new UserNotFoundException(userId));
         marker.setUser(user);
         marker.setContent(request.getContent());
 
@@ -89,7 +91,14 @@ public class MarkerService {
         return marker.getId();
     }
 
-    public void deleteMarker(long markerId) {
+    public void deleteMarker(long userId, long markerId) {
+        Marker marker = markerRepository.findById(markerId)
+                        .orElseThrow(() -> new MarkerNotFoundException(markerId));
+
+        if (marker.getUser().getId() != userId) {
+            throw new AuthorizationException("User is not authorized to delete this marker.")
+        }
+
         markerRepository.deleteById(markerId);
     }
 }
